DB Tools Pinout: signal field width increased with max # of configsets
Signal field width needs to be increased by the amount of characters that are
added for each configset.
Currently an asterisk is added to each signal to indicate it's part of the
first non-default set, two asterisks added to indicate the second set etc.
The fields used do not scroll and need to be wide enough to allow a signal + 1
character possible.