DB Data: SLASEC4, configsets on model AND packages; MSP-EXP430FR2355 LaunchPad
This datasheet specifies configsets based both on models and packages and thus
like a few others use non-hierarchical parents.

Some of the models have a DAC, others don't. Then there is another exception
for the UCB0 ports, which typically support I2C and SPI. On the RSM package
only I2C is supported, which is rather odd really. But it is what it is.
This results in 14 pinsets for 4 packages.

It was a good test for the pinout tool though and exposed another deficiency.

Note that I originally - which is rebased out - had different configsets but I
prefer this arrangement, in which the biggest set (for the FR235x devices) is
the second one. The package-specific ones only have 2 signals, so I placed them
behind using some one-off SQL statements in the file and re-saving.