DB Data: SLASEC4 data, configsets on model AND packages
This datasheet specifies configsets based both on models and packages and thus
like a few others use non-hierarchical parents.

Some of the models have a DAC, others don't. Then there is another exception
for the UCB0 ports, which typically support I2C and SPI. On the RSM package
only I2C is supported, which is rather odd really. But what is, will be.
This results in 14 signalsets for 4 packages.

It was a good test for the pinout tool though and exposed another bug.