DB Tools Pinout: simple ordering/sorting idea to be implemented
This commit has been part of a rebase chain and doesn't really contain much.
It is however nicer to keep it separate as not to dirty other commits with this
specific code.

The basic idea is rather simple:
ordering/sorting is possible on each for each of the packages' pins or simply
based on the default sorting, essentially no ordering by honouring datasheet
index.

Thinking about this ordering functionality also means thinking about what
happens when inserting/deleting and editing. Delete and edit seems to be easy
to implement when the input is reordened, but insertion doesn't. There is no
way of knowing in which datasheet index it should be inserted.
Thus it would probably be best to disable insertion when the pins are ordered
on a package.