DB Tools Pinout: encapsulated view manipulation (indices etc) into PinSetView
Encapsulating this simplifies the code slightly, or at least makes it easier to
read. Technically it doesn't matter a lot, but having easier to read code makes
it simpler to modify the UI based on state etc. Still, this tool isn't an end-
user tool nor is it a product in itself.