DB Data: SLAS704 MSP430FR59[456][789] & x1
Covers MSP-EXP430FR5969 LaunchPad.

Note that this first commit doesn't contain much in the way of pins/signals.
While I was developing the pinout tool I used this datasheet to test the code
because it's one of the most complicating I've seen up to this point.
I noticed some issues with this specific datasheet later on, which were then
corrected in the pinout tool.

The commits have been all rebased after the fixes to the code, so editing them
in order with the current tool is perfectly fine.

Keeping the commits for this datasheet also means that I have a reference on
how to tackle very difficult sheets in the future.