DB Tools: calculating export formatting from data
There are two ways to format data: based on a fixed pattern and based on the
actual data itself. Depending on the data it might be more useful to leave the
formatting defined by the application, to pad no more than required to outline.

The way this calculation works is by iterating over all data of the statement,
calculating widths. Then the statement is reset for actual processing/export.
This means of course twice the amount of work. Considering all data is coming
from an in-memory database with limited amount of records, this shouldn't
create much of a cpu hog.