DB Tools Pinout: [BUGFIX] rare case when pinset used in configset parent only
When a pinset is referenced in a configset and nowhere else (e.g. after edits),
it is possible that it never gets added to the database. This might not be an
actual big problem, but the code does a few safety-checks using asserts to
ensure no unexisting pinsets are referenced.

In this very specific case, an assert was hit. But much rather than hoping it's
not an actual problem, I prefer to ensure correct behaviour.