DB Tools Pinout: [BUGFIX] rare case when pinset used in configset parent only
When a pinset is referenced in a configset and nowhere else (e.g. after edits),
it is possible that it never gets added to the database. This migjt not be an
actual big problem, but the code does a few safety-checks to ensure no pinsets
are referenced that do not exist.

In this very specific case, an assert gets hit. But much rather than hoping
it's not an actual problem, I prefer to ensure correct behaviour.