DB Data: SLAS704 final version, all signals are correct
This was for sure the most difficult pinset to edit, it exposed several bugs
and feature deficiencies in the pinout tool.

The most important to remember in this one is that the fourth pinset is a
specialization of the second one. The third on the other hand is descendant
from the default pinset. Because there are pins shared between pinsets 3 and 4,
which have no connection at all, duplication of these pins is required.
I chose to do it in this particular configuration because it resulted in the
cleanest solution and probably the least duplication.

Essentially this datasheet requires multi-parenting if I want a purely hier-
archical connection, which is currently not supported (and never will be) by
the tool.