DB Tools Pinout: [BUGFIX] DB to view uses all parents of pinsets to assign pins
During conversion from DB objects to view, pins are assigned based on the link
between orderable and pinset. This is correct, but all parent pinsets need to
be considered as well. Previously only 1 parent deep was considered, although
that was sufficient for most datasheets. With datasheets requiring 3 levels
deep it got a bit more complicated.

The VectorIdMapAdapter utility class is NOT a collection in itself - for that
the vector is still used - but it does adapt a vector in an easy way to find
objects based on their 'id' property.

Vector mapping from id is throughout the code base to improve accessed (rather
than looping the entire vector to find a certain key each time). This class
simply centralizes this.