DB Data: SLAS704 added BSL signals
Both UART & I2C are supported in this datasheet, depending on model of course.
But as always, since these signals are fully dependant on the model number
(an added '1' means I2C) there is no need to have them in separate pinsets.

In this particular sheet it would certainly have made things harder as well.