DB Tools Pinout: only allow valid input in datasheet search and error prompt
Previously the user was supposed to enter correct, existing datasheets. Now it
is no longer possible for the function to result in a non-existent datasheet or
model. An error message is shown when it's wrong, prompting the user to correct
the input.

The input is now also automatically upper-cased since the ids for database &
device are upper-case as well. Might just as well make it easier on the user.
That's also the reason why the fields are pre-filled with 'SLAS' for datasheets
and 'MSP430' for models. There is only one - at the time of writing - non-SLAS
datasheet and only the CC430 models wouldn't start with MSP430. It's certain to
be the more user-friendly option.