DB Tools Pinout: uppercase all input (creates a BUG! TODO?)
With the uppercase-conversion code in place, it's easy to implement it on the
signal editing code as well. Now it's impossible to insert non-uppercase sig-
nals, which has been a bit of an issue before while bulk-inserting.

BUG: NEW signals cannot be correctly created anymore because the description is
now also forced to uppercase. This needs to be fixed, but there's another bug
in the description to be fixed as well: the length of the input should be able
to exceed the length of the field.