DB Data: SLAS368 MSP430F15x, MSP430F16x, MSP430F161x
This datasheet isn't directly supported on LaunchPads, but it is the device
that is used as the programming (eZ-FET debug probe/programmer) on the
MSP-EXP430G2 LaunchPad according to the user's guide.
However, while reading the hardware design files' BOM I saw that most Launchpad
boards (including that one ...) use the MSP430F5528 from SLAS590.
So it might be useless in that regard, but it wasn't exactly hard.

A good thing however that came from this one: it seems that the old USART
peripherals before the distinction between USARTA & USARTB weren't one or the
other: the 'classic' USART can handle UART, I2C and SPI on this peripheral!
So it is not possible to simply consider the classic USART as either USARTA/B.